python/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
anyio==3.6.2
    # via httpcore
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
h11==0.14.0
    # via httpcore
httpcore==0.17.0
    # via httpx
httpx==0.24.0
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   httpx
markdown-it-py==2.2.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
pygments==2.15.1
    # via rich
python-frontmatter==1.0.0
    # via -r requirements.in
pyyaml==6.0
    # via python-frontmatter
rich==13.3.5
    # via -r requirements.in
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
yattag==1.15.1
    # via -r requirements.in
